+++ tests/unit_test/test_resource.py
@@ -29,154 +29,218 @@ class TestResource(unittest.TestCase):
 
     def test_valid_args(self):
         ''' Valid arguments. '''
-        resource = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
-                                                   origin=PhyDim2(0, 0)),
+        resource = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                   origin=PhyDim2(0, 0),
+                                                   type=NodeRegion.PROC),
+                            data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                     origin=PhyDim2(0, 0),
+                                                     type=NodeRegion.DATA),
+                                          NodeRegion(dim=PhyDim2(2, 1),
+                                                     origin=PhyDim2(0, 1),
+                                                     type=NodeRegion.DATA)),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
-                            mem_regions=(NodeRegion(dim=PhyDim2(2, 1),
-                                                    origin=PhyDim2(0, 0)),
-                                         NodeRegion(dim=PhyDim2(2, 1),
-                                                    origin=PhyDim2(0, 1)))
                            )
-        self.assertTupleEqual(resource.node_region.dim, (2, 2), 'node_region')
+        self.assertTupleEqual(resource.proc_region.dim, (2, 2), 'proc_region')
         self.assertTupleEqual(resource.dim_array, (16, 16), 'dim_array')
         self.assertEqual(resource.size_gbuf, 131072, 'size_gbuf')
         self.assertEqual(resource.size_regf, 512, 'size_regf')
 
-    def test_invalid_node_region(self):
-        ''' Invalid node_region. '''
-        with self.assertRaisesRegexp(TypeError, 'Resource: .*node_region.*'):
-            _ = Resource(node_region=PhyDim2(2, 2),
+    def test_invalid_proc_region(self):
+        ''' Invalid proc_region. '''
+        with self.assertRaisesRegexp(TypeError, 'Resource: .*proc_region.*'):
+            _ = Resource(proc_region=PhyDim2(2, 2),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=512,
-                         mem_regions=(NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 0)),
-                                      NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 1)))
+                        )
+
+    def test_invalid_proc_region_data(self):
+        ''' Invalid proc_region with type DATA. '''
+        with self.assertRaisesRegexp(ValueError, 'Resource: .*proc_.*type.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.DATA),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
                         )
 
     def test_invalid_dim_array(self):
         ''' Invalid dim_array. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*dim_array.*'):
-            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
-                                                origin=PhyDim2(0, 0)),
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
                          dim_array=(16, 16),
                          size_gbuf=131072,
                          size_regf=512,
-                         mem_regions=(NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 0)),
-                                      NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 1)))
                         )
 
     def test_invalid_size_gbuf(self):
         ''' Invalid size_gbuf. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*size_gbuf.*'):
-            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
-                                                origin=PhyDim2(0, 0)),
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=(131072,),
                          size_regf=512,
-                         mem_regions=(NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 0)),
-                                      NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 1)))
                         )
 
     def test_invalid_size_regf(self):
         ''' Invalid size_regf. '''
         with self.assertRaisesRegexp(TypeError, 'Resource: .*size_regf.*'):
-            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
-                                                origin=PhyDim2(0, 0)),
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
-                         mem_regions=(NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 0)),
-                                      NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 1)))
                         )
 
-    def test_invalid_mem_regions_type(self):
-        ''' Invalid mem_regions type. '''
-        with self.assertRaisesRegexp(TypeError, 'Resource: .*mem_regions.*'):
-            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
-                                                origin=PhyDim2(0, 0)),
+    def test_invalid_data_regions_type(self):
+        ''' Invalid data_regions type. '''
+        with self.assertRaisesRegexp(TypeError, 'Resource: .*data_regions.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=[NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),],
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
-                         mem_regions=[NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 0))]
                         )
-        with self.assertRaisesRegexp(TypeError, 'Resource: .*mem_regions.*'):
-            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
-                                                origin=PhyDim2(0, 0)),
+        with self.assertRaisesRegexp(TypeError, 'Resource: .*data_regions.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=NodeRegion(dim=PhyDim2(2, 1),
+                                                 origin=PhyDim2(0, 0),
+                                                 type=NodeRegion.DATA),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
-                         mem_regions=NodeRegion(dim=PhyDim2(2, 1),
-                                                origin=PhyDim2(0, 0))
                         )
-        with self.assertRaisesRegexp(TypeError, 'Resource: .*mem_regions.*'):
-            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
-                                                origin=PhyDim2(0, 0)),
+        with self.assertRaisesRegexp(TypeError, 'Resource: .*data_regions.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),
+                                       PhyDim2(2, 1)),
                          dim_array=PhyDim2(16, 16),
                          size_gbuf=131072,
                          size_regf=(512,),
-                         mem_regions=(NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 0)),
-                                      PhyDim2(0, 1))
-                        )
-
-    def test_invalid_mem_regions_len(self):
-        ''' Invalid mem_regions len. '''
-        with self.assertRaisesRegexp(ValueError, 'Resource: .*mem_regions.*'):
-            _ = Resource(node_region=NodeRegion(dim=PhyDim2(2, 2),
-                                                origin=PhyDim2(0, 0)),
-                         dim_array=PhyDim2(16, 16),
-                         size_gbuf=131072,
-                         size_regf=(512,),
-                         mem_regions=(NodeRegion(dim=PhyDim2(2, 1),
-                                                 origin=PhyDim2(0, 0)),) * 3
                         )
 
-    def test_mem_region_src(self):
-        ''' Accessor mem_region_src. '''
-        nr1 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 0))
-        nr2 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 1))
-        resource = Resource(node_region=NodeRegion(dim=PhyDim2(4, 4),
-                                                   origin=PhyDim2(0, 0)),
+    def test_invalid_data_regions_len(self):
+        ''' Invalid data_regions len. '''
+        with self.assertRaisesRegexp(ValueError, 'Resource: .*data_regions.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.DATA),) * 3,
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=(512,),
+                        )
+
+    def test_invalid_data_regions_proc(self):
+        ''' Invalid data_regions with type PROC. '''
+        with self.assertRaisesRegexp(ValueError, 'Resource: .*data_.*type.*'):
+            _ = Resource(proc_region=NodeRegion(dim=PhyDim2(2, 2),
+                                                origin=PhyDim2(0, 0),
+                                                type=NodeRegion.PROC),
+                         data_regions=(NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 0),
+                                                  type=NodeRegion.PROC),
+                                       NodeRegion(dim=PhyDim2(2, 1),
+                                                  origin=PhyDim2(0, 1),
+                                                  type=NodeRegion.DATA)),
+                         dim_array=PhyDim2(16, 16),
+                         size_gbuf=131072,
+                         size_regf=512,
+                        )
+
+    def test_src_data_region(self):
+        ''' Accessor src_data_region. '''
+        nr1 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 0),
+                         type=NodeRegion.DATA)
+        nr2 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 1),
+                         type=NodeRegion.DATA)
+        resource = Resource(proc_region=NodeRegion(dim=PhyDim2(4, 4),
+                                                   origin=PhyDim2(0, 0),
+                                                   type=NodeRegion.PROC),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
-                            mem_regions=(nr1, nr2))
-        self.assertEqual(resource.mem_region_src(), nr1, 'mem_region_src')
-        resource = Resource(node_region=NodeRegion(dim=PhyDim2(4, 4),
-                                                   origin=PhyDim2(0, 0)),
+                            data_regions=(nr1, nr2))
+        self.assertEqual(resource.src_data_region(), nr1, 'src_data_region')
+        resource = Resource(proc_region=NodeRegion(dim=PhyDim2(4, 4),
+                                                   origin=PhyDim2(0, 0),
+                                                   type=NodeRegion.PROC),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
-                            mem_regions=(nr1,))
-        self.assertEqual(resource.mem_region_src(), nr1, 'mem_region_src')
+                            data_regions=(nr1,))
+        self.assertEqual(resource.src_data_region(), nr1, 'src_data_region')
 
-    def test_mem_region_dst(self):
-        ''' Accessor mem_region_dst. '''
-        nr1 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 0))
-        nr2 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 1))
-        resource = Resource(node_region=NodeRegion(dim=PhyDim2(4, 4),
-                                                   origin=PhyDim2(0, 0)),
+    def test_dst_data_region(self):
+        ''' Accessor dst_data_region. '''
+        nr1 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 0),
+                         type=NodeRegion.DATA)
+        nr2 = NodeRegion(dim=PhyDim2(2, 1), origin=PhyDim2(0, 1),
+                         type=NodeRegion.DATA)
+        resource = Resource(proc_region=NodeRegion(dim=PhyDim2(4, 4),
+                                                   origin=PhyDim2(0, 0),
+                                                   type=NodeRegion.PROC),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
-                            mem_regions=(nr1, nr2))
-        self.assertEqual(resource.mem_region_dst(), nr2, 'mem_region_dst')
-        resource = Resource(node_region=NodeRegion(dim=PhyDim2(4, 4),
-                                                   origin=PhyDim2(0, 0)),
+                            data_regions=(nr1, nr2))
+        self.assertEqual(resource.dst_data_region(), nr2, 'dst_data_region')
+        resource = Resource(proc_region=NodeRegion(dim=PhyDim2(4, 4),
+                                                   origin=PhyDim2(0, 0),
+                                                   type=NodeRegion.PROC),
                             dim_array=PhyDim2(16, 16),
                             size_gbuf=131072,
                             size_regf=512,
-                            mem_regions=(nr1,))
-        self.assertEqual(resource.mem_region_dst(), nr1, 'mem_region_dst')
+                            data_regions=(nr1,))
+        self.assertEqual(resource.dst_data_region(), nr1, 'dst_data_region')